+++ test/persistence/test_DummyMongoDbConnection.py
@@ -6,7 +6,7 @@ from pip_services3_commons.config.ConfigParams import ConfigParams
 from pip_services3_commons.refer.Descriptor import Descriptor
 from pip_services3_commons.refer.References import References
 
-from pip_services3_mongodb.persistence.MongoDbConnection import MongoDbConnection
+from pip_services3_mongodb.connect.MongoDbConnection import MongoDbConnection
 from .DummyMongoDbPersistence import DummyMongoDbPersistence
 from ..DummyPersistenceFixture import DummyPersistenceFixture