+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> MongoDB components for Python Changelog
 
+## <a name="3.1.0"></a> 3.1.0 (2021-05-12)
+
+### Bug Fixes
+* fix docs
+* fix private, protected method names
+
+### Features
+* Moved MongoDbConnection to connect package
+* To IdentifiableMongoDbPersistence added _convert_from_public_partial
+* Added tracing for methods
+* Added MongoDbIndex
+* Added type hints
+
+
 ## <a name="3.0.6"></a> 3.0.6 (2021-04-22)
 
 ### Bug Fixes